[cxMonitor] Move to cxMonitor to own file [cxMonitor] Take pending pool size limit as a parameter from config [cxMonitor] Added prints for current status
